refactor: move state and queue from generic creator implementation into counter usecase
