+++ src/creator/mod.rs
@@ -1,15 +1,5 @@
 // Core traits and types
 pub mod core;
 
-// Generic creator implementations
-pub mod default_creator;
-pub mod listening_creator;
-
-// Tests
-#[cfg(test)]
-pub mod tests;
-
 // Re-export main types for easy access
-pub use core::{BoxedCreator, CreatorConfig, SimpleTaskQueue};
-pub use default_creator::DefaultCreator;
-pub use listening_creator::ListeningCreator;
+pub use core::BoxedCreator;
