+++ src/usecases/counter/factories.rs
@@ -1,29 +1,20 @@
 use anyhow::Result;
-use async_trait::async_trait;
 
 use super::types::DefaultTaskData;
-use crate::creator::core::{DefaultState, TaskDataFactory, TaskRequest};
-
-// Use DefaultState directly instead of type alias
-type CounterState = DefaultState<u64>;
+use crate::creator::core::TaskRequest;
 
 /// Default task data factory for counter operations
 pub struct DefaultTaskDataFactory;
 
-#[async_trait]
-impl TaskDataFactory for DefaultTaskDataFactory {
-    type State = CounterState;
-    type TaskData = DefaultTaskData;
-
-    async fn create_task_data(&self, _state: &Self::State) -> Result<Self::TaskData> {
+impl DefaultTaskDataFactory {
+    pub async fn create_task_data(&self) -> Result<DefaultTaskData> {
         Ok(DefaultTaskData::default())
     }
 
-    async fn create_task_data_from_request(
+    pub async fn create_task_data_from_request(
         &self,
-        _state: &Self::State,
         request: &TaskRequest,
-    ) -> Result<Self::TaskData> {
+    ) -> Result<DefaultTaskData> {
         Ok(DefaultTaskData {
             var1: request.body.var1.clone(),
             var2: request.body.var2.clone(),
